Refacterd the code and rearranged the file structu-re
