+++ .gitignore
@@ -31,3 +31,4 @@ build/
 
 ### VS Code ###
 .vscode/
+node_modules
