+++ Frontend/app/(AuthScreens)/Login-Screen/ForgotPasswordScreen.js
@@ -1,0 +1,85 @@
+import React, { useState } from 'react';
+import { View, TextInput, Text, TouchableOpacity, Alert, StyleSheet } from 'react-native';
+import { useNavigation } from '@react-navigation/native';
+
+const ForgotPasswordScreen = () => {
+  const navigation = useNavigation();
+  const [email, setEmail] = useState('');
+
+  const handleResetPassword = () => {
+    if (!email) {
+      Alert.alert('Error', 'Please enter your email');
+      return;
+    }
+
+    Alert.alert('Success', 'Password reset link sent!');
+    navigation.goBack();
+  };
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Forgot Password</Text>
+
+      <TextInput
+        placeholder="Enter your email"
+        style={styles.input}
+        keyboardType="email-address"
+        autoCapitalize="none"
+        value={email}
+        onChangeText={setEmail}
+      />
+
+      <TouchableOpacity style={styles.button} onPress={handleResetPassword}>
+        <Text style={styles.buttonText}>Send Reset Link</Text>
+      </TouchableOpacity>
+
+      <TouchableOpacity onPress={() => navigation.goBack()}>
+        <Text style={styles.goBack}>← Go Back</Text>
+      </TouchableOpacity>
+    </View>
+  );
+};
+
+export default ForgotPasswordScreen;
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    padding: 24,
+    justifyContent: 'center',
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: '700',
+    marginBottom: 24,
+    textAlign: 'center',
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: '#ccc',
+    borderRadius: 10,
+    paddingHorizontal: 16,
+    paddingVertical: 14,
+    marginBottom: 16,
+    fontSize: 16,
+  },
+  button: {
+    backgroundColor: '#FF9500',
+    paddingVertical: 14,
+    borderRadius: 10,
+    marginTop: 10,
+  },
+  buttonText: {
+    color: '#fff',
+    fontWeight: '600',
+    fontSize: 16,
+    textAlign: 'center',
+  },
+  goBack: {
+    marginTop: 20,
+    color: '#007AFF',
+    textAlign: 'center',
+    fontSize: 16,
+  },
+});
