+++ Frontend/app/(AuthScreens)/Login-Screen/LoginFormScreen.js
@@ -1,0 +1,102 @@
+import React, { useState } from 'react';
+import { View, Text, TextInput, TouchableOpacity, Alert, StyleSheet } from 'react-native';
+import { useNavigation } from '@react-navigation/native';
+
+const LoginFormScreen = () => {
+  const navigation = useNavigation();
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+   const [errorMessage, setErrorMessage] = useState('');
+  
+
+  const handleLogin = () => {
+    if (!email || !password) {
+      setErrorMessage('All fields are required');
+      return;
+    }
+
+    if (password.length < 6) {
+      setErrorMessage('Error', 'Password must be at least 6 characters');
+      return;
+    }
+
+    // Fake login logic
+    Alert.alert('Success', 'Logged in successfully!');
+    navigation.replace('HomeScreen'); // Replace with HomeScreen
+  };
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.header}>Login</Text>
+
+      <TextInput
+        style={styles.input}
+        placeholder="Email"
+        placeholderTextColor="#aaa"
+        keyboardType="email-address"
+        value={email}
+        onChangeText={setEmail}
+      />
+
+      <TextInput
+        style={styles.input}
+        placeholder="Password"
+        placeholderTextColor="#aaa"
+        secureTextEntry
+        value={password}
+        onChangeText={setPassword}
+      />
+
+      <TouchableOpacity style={styles.loginBtn} onPress={handleLogin}>
+        <Text style={styles.loginText}>Login</Text>
+      </TouchableOpacity>
+
+      <TouchableOpacity onPress={() => navigation.goBack()}>
+        <Text style={styles.backText}>← Go Back</Text>
+      </TouchableOpacity>
+    </View>
+  );
+};
+
+export default LoginFormScreen;
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    padding: 24,
+    justifyContent: 'center',
+  },
+  header: {
+    fontSize: 22,
+    fontWeight: '700',
+    textAlign: 'center',
+    marginBottom: 24,
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: '#ccc',
+    padding: 12,
+    borderRadius: 8,
+    marginBottom: 16,
+    fontSize: 16,
+  },
+  loginBtn: {
+    backgroundColor: '#007AFF',
+    padding: 14,
+    borderRadius: 8,
+    alignItems: 'center',
+    marginTop: 12,
+  },
+  loginText: {
+    color: '#fff',
+    fontSize: 16,
+    fontWeight: '600',
+  },
+  backText: {
+    marginTop: 20,
+    color: '#007AFF',
+    fontSize: 15,
+    textAlign: 'center',
+  },
+});
