--- Login-Screen/HomeScreen.js
@@ -1,32 +1,0 @@
-import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
-
-const HomeScreen = () => {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Welcome to Bisa HomeScreen!</Text>
-      <Text style={styles.subtitle}>You are now logged in.</Text>
-    </View>
-  );
-};
-
-export default HomeScreen;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    padding: 24,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: '700',
-    marginBottom: 8,
-  },
-  subtitle: {
-    fontSize: 18,
-    color: '#555',
-  },
-});
