+++ package.json
@@ -1,0 +1,9 @@
+{
+  "devDependencies": {
+    "@types/react-native": "^0.72.8",
+    "typescript": "~5.8.3"
+  },
+  "dependencies": {
+    "@expo/vector-icons": "^14.1.0"
+  }
+}
